+++ bbruntime_dll/bbruntime_dll.cpp
@@ -199,7 +199,7 @@ static void link(){
 
 		if( !isalnum(t[0]) ) t=t.substr(1);
 
-		for( int k=0;k<t.size();++k ){
+		for( int k=0;k<(int)t.size();++k ){
 			if( isalnum(t[k]) || t[k]=='_' ) continue;
 			t=t.substr( 0,k );break;
 		}
